Update find_n-th_ugly_number.py
This python code takes input from the user and finds the n-th number in the Ugly Number Series. At the end of the code there is a mock testing section given with sample value to test out the code. The code can be modified to take input directly from the user terminal by replacing the mock testing section with some input/output lines, keeping the algorithm of the rest of the code same.